Documentation
- Wrote documentation for code base
