--- app/src/main/java/com/m/ammar/itaskmanager/data/SnackbarEvent.kt
@@ -1,15 +1,0 @@
-package com.m.ammar.itaskmanager.data
-
-import com.m.ammar.itaskmanager.data.local.model.Task
-
-sealed class SnackbarEvent {
-    data class TaskDeleted(
-        val task: Task,
-        val undo: () -> Unit  // Now properly defined
-    ) : SnackbarEvent()
-
-    data class TaskCompleted(
-        val task: Task,
-        val undo: () -> Unit  // Now properly defined
-    ) : SnackbarEvent()
-}
