+++ app/src/main/java/com/m/ammar/itaskmanager/data/dihilt/DatabaseModule.kt
@@ -11,10 +11,19 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
 
+/**
+ * Dagger Hilt module that provides database-related dependencies.
+ */
 @Module
 @InstallIn(SingletonComponent::class)
 object DatabaseModule {
 
+    /**
+     * Provides a singleton instance of [AppDatabase] using Room.
+     *
+     * @param context Application context injected by Hilt.
+     * @return The Room database instance.
+     */
     @Provides
     @Singleton
     fun provideAppDatabase(
@@ -27,6 +36,12 @@ object DatabaseModule {
         ).build()
     }
 
+    /**
+     * Provides the [TaskDao] instance to perform task-related DB operations.
+     *
+     * @param database The [AppDatabase] instance.
+     * @return The DAO for tasks.
+     */
     @Provides
     fun provideTaskDao(database: AppDatabase): TaskDao {
         return database.taskDao()
