+++ app/src/main/java/com/m/ammar/itaskmanager/data/local/model/Task.kt
@@ -1,9 +1,18 @@
 package com.m.ammar.itaskmanager.data.local.model
 
-import android.opengl.Visibility
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
+/**
+ * Data class representing a task in the application.
+ *
+ * @property id Unique identifier for the task. Automatically generated by Room.
+ * @property title The title of the task.
+ * @property description A brief description of the task. Can be null.
+ * @property priority The priority level of the task (HIGH, MEDIUM, or LOW).
+ * @property dueDate The due date of the task represented as a Unix timestamp (milliseconds).
+ * @property isCompleted A flag indicating whether the task is completed. Defaults to false.
+ */
 @Entity(tableName = "tasks")
 data class Task(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
