+++ app/src/main/java/com/m/ammar/itaskmanager/navigation/AppNavigation.kt
@@ -1,54 +1,57 @@
 package com.m.ammar.itaskmanager.navigation
 
-import androidx.compose.foundation.layout.Box
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.SnackbarDuration
-import androidx.compose.material3.SnackbarHost
-import androidx.compose.material3.SnackbarHostState
-import androidx.compose.material3.SnackbarResult
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import com.m.ammar.itaskmanager.data.SnackbarEvent
 import com.m.ammar.itaskmanager.data.local.model.Task
-import com.m.ammar.itaskmanager.ui.task_detail.TaskDetailsScreen
-import com.m.ammar.itaskmanager.ui.task_creation.TaskCreationScreen
 import com.m.ammar.itaskmanager.ui.home.HomeScreen
 import com.m.ammar.itaskmanager.ui.home.HomeViewModel
 import com.m.ammar.itaskmanager.ui.settings.SettingsScreen
 import com.m.ammar.itaskmanager.ui.settings.SettingsViewModel
+import com.m.ammar.itaskmanager.ui.task_creation.TaskCreationScreen
+import com.m.ammar.itaskmanager.ui.task_detail.TaskDetailsScreen
 
+/**
+ * Composable function that handles the app's navigation.
+ * Sets up navigation and routing between different screens of the app.
+ *
+ * @param modifier Modifier to customize the layout.
+ * @param navController The navigation controller used for managing navigation actions.
+ */
 @Composable
 fun AppNavigation(
     modifier: Modifier = Modifier,
     navController: NavHostController,
 ) {
+    // Initialize ViewModels using Hilt
     val homeViewModel: HomeViewModel = hiltViewModel()
     val settingsViewModel: SettingsViewModel = hiltViewModel()
 
+    // Define the navigation host with routes to all major destinations
     NavHost(
         modifier = modifier,
         navController = navController,
         startDestination = TopLevelDestination.Home.route
     ) {
+        // Composable for the Home screen
         composable(route = TopLevelDestination.Home.route) {
+            // Collect current data from HomeViewModel using lifecycle-aware state
             val tasks by homeViewModel.tasksFlow.collectAsStateWithLifecycle()
             val currentSort by homeViewModel.sortOption.collectAsStateWithLifecycle()
             val currentFilter by homeViewModel.filterOption.collectAsStateWithLifecycle()
 
+            // Load data when the Home screen is first launched
             LaunchedEffect(Unit) {
                 homeViewModel.loadData()
             }
 
-
+            // Display the HomeScreen composable
             HomeScreen(
                 tasks = tasks,
                 sortOption = currentSort,
@@ -66,9 +69,9 @@ fun AppNavigation(
                 onUndoDeleted = { homeViewModel.undoDelete() },
                 onSettingsClick = { navController.navigate(TopLevelDestination.Settings.route) }
             )
-
         }
 
+        // Composable for the Task Creation screen
         composable(route = TopLevelDestination.CreateTask.route) {
             TaskCreationScreen(
                 onSaveClick = { title, description, priority, dueDate ->
@@ -87,9 +90,12 @@ fun AppNavigation(
             )
         }
 
+        // Composable for the Task Detail screen
         composable(route = TopLevelDestination.TaskDetail.route) {
+            // Collect selected task from HomeViewModel
             val task by homeViewModel.selectedTask.collectAsStateWithLifecycle()
 
+            // Display the TaskDetailsScreen if a task is selected
             task?.let { nonNullTask ->
                 TaskDetailsScreen(
                     task = nonNullTask,
@@ -105,6 +111,7 @@ fun AppNavigation(
             }
         }
 
+        // Composable for the Settings screen
         composable(route = TopLevelDestination.Settings.route) {
             SettingsScreen(
                 themeMode = settingsViewModel.themeMode.value,
