--- app/src/main/java/com/m/ammar/itaskmanager/ui/components/Components.kt
@@ -1,52 +1,0 @@
-package com.m.ammar.itaskmanager.ui.components
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.size
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
-
-
-@Composable
-fun VerticalSpacer(size: Int) = Spacer(modifier = Modifier.height(size.dp))
-
-@Composable
-fun AppLoader(
-    modifier: Modifier = Modifier,
-    size: Dp = 40.dp,
-    color: Color = MaterialTheme.colorScheme.primary,
-    strokeWidth: Dp = 4.dp
-) {
-    Box(modifier = modifier) {
-        CircularProgressIndicator(
-            modifier = Modifier
-                .size(size)
-                .align(Alignment.Center),
-            color = color,
-            strokeWidth = strokeWidth
-        )
-    }
-}
-
-@Composable
-fun ErrorItem(
-    text: String,
-    modifier: Modifier = Modifier,
-    color: Color = MaterialTheme.colorScheme.error
-) {
-    Box(modifier = modifier) {
-        Text(
-            text = text,
-            modifier = Modifier.align(Alignment.Center),
-            color = color
-        )
-    }
-}
