+++ app/src/main/java/com/m/ammar/itaskmanager/ui/home/HomeViewModel.kt
@@ -2,7 +2,6 @@ package com.m.ammar.itaskmanager.ui.home
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.m.ammar.itaskmanager.data.SnackbarEvent
 import com.m.ammar.itaskmanager.data.local.enums.FilterOption
 import com.m.ammar.itaskmanager.data.local.enums.SortOption
 import com.m.ammar.itaskmanager.data.local.dao.TaskDao
@@ -22,6 +21,9 @@ class HomeViewModel @Inject constructor(
     private val taskDao: TaskDao,
 ) : ViewModel() {
 
+    /**
+     * Loads all tasks from the database and updates the state.
+     */
     fun loadData() {
         viewModelScope.launch {
             val tasks = taskDao.getTasksSortedAlphabetically()
@@ -29,6 +31,11 @@ class HomeViewModel @Inject constructor(
         }
     }
 
+    /**
+     * Adds a new task to the database and reloads the task list.
+     *
+     * @param task The task to be added.
+     */
     fun addTask(task: Task) {
         viewModelScope.launch {
             taskDao.insertTask(task)
@@ -37,6 +44,12 @@ class HomeViewModel @Inject constructor(
     }
 
     private var lastDeleteTask: ArrayList<Task?> = arrayListOf()
+
+    /**
+     * Deletes a task from the database and stores it for possible undo.
+     *
+     * @param task The task to be deleted.
+     */
     fun deleteTask(task: Task) {
         viewModelScope.launch {
             lastDeleteTask.add(task)
@@ -45,6 +58,11 @@ class HomeViewModel @Inject constructor(
         }
     }
 
+    /**
+     * Updates a task in the database and reloads the task list.
+     *
+     * @param task The task to be updated.
+     */
     fun updateTask(task: Task) {
         viewModelScope.launch {
             taskDao.updateTask(task)
@@ -58,9 +76,11 @@ class HomeViewModel @Inject constructor(
     private val _filterOption = MutableStateFlow(FilterOption.ALL)
     val filterOption = _filterOption.asStateFlow() // Now reactive
 
-
     private val _allTasks = MutableStateFlow<List<Task>>(emptyList())
 
+    /**
+     * Combines the task list with the selected sorting and filtering options to return a reactive flow.
+     */
     val tasksFlow = combine(
         _allTasks,
         _sortOption,
@@ -74,18 +94,25 @@ class HomeViewModel @Inject constructor(
         viewModelScope,
         SharingStarted.WhileSubscribed(50000),
         emptyList()
-
-
     )
 
+    /**
+     * Sets the sorting option for tasks.
+     *
+     * @param option The sort option to be set.
+     */
     fun setSortOption(option: SortOption) {
         _sortOption.value = option
     }
 
+    /**
+     * Sets the filter option for tasks.
+     *
+     * @param option The filter option to be set.
+     */
     fun setFilterOption(option: FilterOption) {
         _filterOption.value = option
     }
-
 
     private fun List<Task>.applySort(sort: SortOption): List<Task> = when (sort) {
         SortOption.DUE_DATE -> sortedBy { it.dueDate }
@@ -102,10 +129,20 @@ class HomeViewModel @Inject constructor(
     private val _selectedTask = MutableStateFlow<Task?>(null)
     val selectedTask: StateFlow<Task?> = _selectedTask
 
+    /**
+     * Sets the selected task.
+     *
+     * @param task The task to be selected.
+     */
     fun selectTask(task: Task) {
         _selectedTask.value = task
     }
 
+    /**
+     * Marks a task as completed and reloads the task list.
+     *
+     * @param it The task to be completed.
+     */
     fun completeTask(it: Task) {
         viewModelScope.launch {
             taskDao.updateTask(it.copy(isCompleted = true))
@@ -113,6 +150,9 @@ class HomeViewModel @Inject constructor(
         }
     }
 
+    /**
+     * Restores the last deleted task.
+     */
     fun undoDelete() {
         viewModelScope.launch {
             lastDeleteTask.lastOrNull()?.let {
