+++ app/src/main/java/com/m/ammar/itaskmanager/ui/settings/SettingsViewModel.kt
@@ -18,23 +18,31 @@ import javax.inject.Inject
 @HiltViewModel
 class SettingsViewModel @Inject constructor(
     val settingsRepository: SettingsRepository
-) :
-    ViewModel() {
+) : ViewModel() {
 
+    /**
+     * Mutable state holding the current theme mode (System, Light, or Dark).
+     * Exposed as immutable state via [themeMode].
+     */
     private val _themeMode = mutableStateOf(ThemeMode.SYSTEM)
     val themeMode: State<ThemeMode> get() = _themeMode
 
+    /**
+     * Mutable state holding the currently selected color set.
+     * Exposed as immutable state via [selectedColorSet].
+     */
     private val _selectedColorSet = mutableStateOf(ThemeSet1)
     val selectedColorSet: State<ThemeSet> get() = _selectedColorSet
 
-        init {
-//    fun loadTheme() {
+    init {
+        // Collects the theme mode value from the repository and updates the UI state
         viewModelScope.launch {
             settingsRepository.themeMode.collect { mode ->
                 _themeMode.value = ThemeMode.valueOf(mode)
             }
         }
 
+        // Collects the selected theme set from the repository and updates the UI state
         viewModelScope.launch {
             settingsRepository.selectedThemeSet.collect { themeSetName ->
                 _selectedColorSet.value = when (themeSetName) {
@@ -48,6 +56,11 @@ class SettingsViewModel @Inject constructor(
         }
     }
 
+    /**
+     * Updates the theme mode in both the view model state and the repository.
+     *
+     * @param mode The new theme mode to set (e.g., Light, Dark, System).
+     */
     fun updateThemeMode(mode: ThemeMode) {
         viewModelScope.launch {
             _themeMode.value = mode
@@ -55,6 +68,11 @@ class SettingsViewModel @Inject constructor(
         }
     }
 
+    /**
+     * Updates the selected color set in both the view model state and the repository.
+     *
+     * @param colorSet The new color set to apply (e.g., ThemeSet1, ThemeSet2).
+     */
     fun updateColorSet(colorSet: ThemeSet) {
         viewModelScope.launch {
             _selectedColorSet.value = colorSet
